requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
pyncclient==0.7
python-dotenv==0.21.0
requests==2.28.1
sharepy==2.0.0
six==1.16.0
urllib3==1.26.13
